Record removido para se adequar a arquitetura
