+++ src/main/java/com/epcafes/DAO/FuncionarioRepository.java
@@ -1,6 +1,6 @@
 package com.epcafes.DAO;
 
-import com.epcafes.model.funcionario.Funcionario;
+import com.epcafes.modelo.funcionario.Funcionario;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface FuncionarioRepository extends JpaRepository<Funcionario, Long> {
