--- src/main/java/com/epcafes/model/funcionario/DadosCadastroFuncionario.java
@@ -1,4 +1,0 @@
-package com.epcafes.model.funcionario;
-
-public record DadosCadastroFuncionario(String nome, Double salario, String nascimento) {
-}
