+++ src/main/java/com/epcafes/modelo/funcionario/DadosAlteraFuncionario.java
@@ -1,4 +1,4 @@
-package com.epcafes.model.funcionario;
+package com.epcafes.modelo.funcionario;
 
 public record DadosAlteraFuncionario(Long id, String nome, Double salario, String nascimento) {
