+++ src/main/java/com/epcafes/service/FuncionarioService.java
@@ -1,6 +1,6 @@
 package com.epcafes.service;
 
-import com.epcafes.model.funcionario.Funcionario;
+import com.epcafes.modelo.funcionario.Funcionario;
 import com.epcafes.DAO.FuncionarioRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
@@ -10,21 +10,21 @@ import java.util.List;
 @Service
 public class FuncionarioService {
     @Autowired
-    private FuncionarioRepository repositorio;
+    private FuncionarioRepository funcionarioRepository;
 
     public void salvar(Funcionario funcionario){
-        repositorio.save(funcionario);
+        funcionarioRepository.save(funcionario);
     }
 
     public List<Funcionario> acharTodos(){
-        return repositorio.findAll();
+        return funcionarioRepository.findAll();
     }
 
     public Funcionario acharPorID(Long id){
-        return repositorio.getReferenceById(id);
+        return funcionarioRepository.getReferenceById(id);
     }
 
     public void deletarPorId(Long id){
-        repositorio.deleteById(id);
+        funcionarioRepository.deleteById(id);
     }
 }
